+++ CHANGELOG.md
@@ -27,6 +27,13 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   their responses
   ([discussion](https://github.com/CBIIT/ccdi-federation-api/discussions/79),
   [#95](https://github.com/CBIIT/ccdi-federation-api/pull/95)).
+- Adds Library Source Material ([link to
+  discussion](https://github.com/CBIIT/ccdi-federation-api/discussions/119),
+  [#118](https://github.com/CBIIT/ccdi-federation-api/pull/118))
+- Adds Specimen Molecular Analyte Type ([link to
+  discussion](https://github.com/CBIIT/ccdi-federation-api/discussions/116),
+  [#123](https://github.com/CBIIT/ccdi-federation-api/pull/123),
+  [#122](https://github.com/CBIIT/ccdi-federation-api/issues/122))
 
 ### Revised